+++ src/lib/systemPoller.js
@@ -101,34 +101,34 @@ function applyConfig(originalConfig) {
     const currPollerIDs = module.exports.getPollerTimers();
 
     systemPollers.forEach((pollerConfig) => {
-        // TODO: keep using traceName as ID for now
-        // tackle using IDs with AUTOTOOL-1834
-        // and see if there's a simple way to skip 'update' altogether
-        // if the exact same poller is configured and already running
-
-        newPollerIDs.push(pollerConfig.traceName);
-        pollerConfig.tracer = util.tracer.createFromConfig(
-            TRACER_CLASS_NAME, pollerConfig.traceName, pollerConfig
-        );
-        const baseMsg = `system poller ${pollerConfig.traceName}. Interval = ${pollerConfig.interval} sec.`;
-        // add to data context to track source poller config and destination(s)
-        pollerConfig.destinationIds = originalConfig.mappings[pollerConfig.id];
-        if (pollerConfig.interval === 0) {
-            logger.info(`Configuring non-polling ${baseMsg}`);
-            if (currPollerIDs[pollerConfig.traceName]) {
-                util.stop(currPollerIDs[pollerConfig.traceName]);
-            }
-            currPollerIDs[pollerConfig.traceName] = undefined;
-        } else if (currPollerIDs[pollerConfig.traceName]) {
-            logger.info(`Updating ${baseMsg}`);
-            currPollerIDs[pollerConfig.traceName] = util.update(
-                currPollerIDs[pollerConfig.traceName], safeProcess, pollerConfig, pollerConfig.interval
-            );
+        const existingPoller = currPollerIDs[pollerConfig.traceName];
+        if (pollerConfig.skipUpdate && existingPoller) {
+            newPollerIDs.push(pollerConfig.traceName);
         } else {
-            logger.info(`Starting ${baseMsg}`);
-            currPollerIDs[pollerConfig.traceName] = util.start(
-                safeProcess, pollerConfig, pollerConfig.interval
+            newPollerIDs.push(pollerConfig.traceName);
+            pollerConfig.tracer = util.tracer.createFromConfig(
+                TRACER_CLASS_NAME, pollerConfig.traceName, pollerConfig
             );
+            const baseMsg = `system poller ${pollerConfig.traceName}. Interval = ${pollerConfig.interval} sec.`;
+            // add to data context to track source poller config and destination(s)
+            pollerConfig.destinationIds = originalConfig.mappings[pollerConfig.id];
+            if (pollerConfig.interval === 0) {
+                logger.info(`Configuring non-polling ${baseMsg}`);
+                if (currPollerIDs[pollerConfig.traceName]) {
+                    util.stop(currPollerIDs[pollerConfig.traceName]);
+                }
+                currPollerIDs[pollerConfig.traceName] = undefined;
+            } else if (currPollerIDs[pollerConfig.traceName]) {
+                logger.info(`Updating ${baseMsg}`);
+                currPollerIDs[pollerConfig.traceName] = util.update(
+                    currPollerIDs[pollerConfig.traceName], safeProcess, pollerConfig, pollerConfig.interval
+                );
+            } else {
+                logger.info(`Starting ${baseMsg}`);
+                currPollerIDs[pollerConfig.traceName] = util.start(
+                    safeProcess, pollerConfig, pollerConfig.interval
+                );
+            }
         }
     });